requirements.txt
ansible-runner==2.3.6
docutils==0.21.2
lockfile==0.12.2
packaging==24.0
pexpect==4.9.0
ptyprocess==0.7.0
python-daemon==3.0.1
bcrypt==4.1.3
cffi==1.16.0
cryptography==42.0.5
gitdb==4.0.11
GitPython==3.1.43
paramiko==3.4.0
pycparser==2.22
PyNaCl==1.5.0
PyYAML==6.0.1
six==1.16.0
smmap==5.0.1
